requirements.txt
fastapi
uvicorn
torch
langchain
langchain_community
qdrant_client
langchain_qdrant
fastembed
python-dotenv
python-multipart
langchainhub
langchain_groq
PyMuPDF
gradio
rank_bm25
langgraph
bm25
scikit-learn